Move MobNBT
