+++ src/main/java/uhc/data/nbt/entity/mobs/PlayerNBT.java
@@ -1,6 +1,5 @@
 package uhc.data.nbt.entity.mobs;
 
-import uhc.data.nbt.entity.MobNBT;
 import uhc.data.nbt.tags.*;
 import uhc.game.GameModeId;
 import uhc.resource.dimension.DimensionId;
